requirement.txt
streamlit==1.35.0
langchain==0.2.14
langchain-core==0.2.39
langchain-community==0.2.12
pandas==2.0.2
langchain-groq==0.1.9
python-dotenv

